omni: overlay: set config_overrideHasPermanentMenuKey default to no key
dont rely on the autodetect method cause it relies
on !wm.hasNavigationBar() but there are hw button devices
that dont have a menu button.

Basically this API is only used from ancient apps that use
it to check if a menu is shown or not. But we always want
them to force to show it to prevents problems

Change-Id: I8976f772d5801612bac3bdddeecc314560b51c4e

Conflicts:
	overlay/common/frameworks/base/core/res/res/values/config.xml